fix: classify unusable worktrees from local checkout only
Do not treat external fetch/ssh errors that mention "no such file or
directory" as recreate-worthy checkouts — force CleanupWorktree would
destroy interrupted agent dirt. Determine unusability from path presence
or local git-tree integrity errors, cover the real gateway prepare
lifecycle, and move dirty-adopt recovery tests into a dedicated file
with shared fixtures so runner_test.go stays under the growth threshold.

Generated-By: looper 0.11.0 (runner=fixer, agent=grok-build)
